code cleanup: calibration details binning as int, not double (#1398)
The binning value can be stored as an int, not a double.

This is backwards compatible with stored calibration values
as the stored values can be read back as integer values
even though they were written out as doubles.

+++ src/scope.cpp
@@ -1120,7 +1120,7 @@ void Scope::SetCalibration(const Calibration& cal)
     Mount::SetCalibration(cal);
 }
 
-void Scope::SetCalibrationDetails(const CalibrationDetails& calDetails, double xAngle, double yAngle, double binning)
+void Scope::SetCalibrationDetails(const CalibrationDetails& calDetails, double xAngle, double yAngle, int binning)
 {
     m_calibrationDetails = calDetails;
